how delete picture of table

+++ app/Http/Controllers/EmpleadoController.php
@@ -87,7 +87,11 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
+        $empleado = Empleado::findOrFail($id);
+        if (Storage::delete('public/' . $empleado->Foto)) {
+            Empleado::destroy($id);
+        }
+
         return redirect('empleado');
     }
 }
